requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.0.0
black==24.10.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1